唉, 重构
