+++ src/callbacks/friendship.py
@@ -3,7 +3,7 @@ import random
 from telegram import Update
 from telegram.ext import ContextTypes
 
-from ..data import ohayo_word, oyasumi_word
+from ..common.utils import ohayo_word, oyasumi_word
 from ..logger import logger
 from .jobs import send_message
