+++ src/common/utils.py
@@ -1,14 +1,17 @@
+import os
+import pathlib
 import random
 import re
 from operator import attrgetter
+import glob
+from ..logger import logger
+import json
 
 from telegram import (
     InlineKeyboardButton,
     InlineKeyboardMarkup,
 )
-from telegram.constants import ChatID
 
-fake_users_id = [ChatID.FAKE_CHANNEL, ChatID.ANONYMOUS_ADMIN, ChatID.SERVICE_CHAT]
 
 back_home_markup = InlineKeyboardMarkup(
     [
@@ -48,3 +51,45 @@ def parse_arguments(text: str) -> list[str]:
     parsed_arguments = [group[0] or group[1] for group in arguments]
 
     return parsed_arguments
+
+
+def _load_word_dict():
+    word_dict_path_internal = (
+        pathlib.Path(__file__).parent.parent.parent / "resource" / "word_dicts"
+    )
+    word_dict_path_user = (
+        pathlib.Path(__file__).parent.parent.parent / "data" / "word_dicts"
+    )
+    word_dict = {}
+    for file in glob.glob(f"{word_dict_path_internal}" + r"/*.json"):
+        logger.debug(f"加载内置词库: {file}")
+        try:
+            with open(file, "r") as f:
+                for k, v in json.load(f).items():
+                    if k in word_dict:
+                        word_dict[k].extend(v)
+                    else:
+                        word_dict[k] = v
+        except Exception as e:
+            logger.error(f"加载词库失败: {file}: {e.__class__.__name__}: {e}")
+            continue
+    if os.path.exists(word_dict_path_user):
+        for file in glob.glob(f"{word_dict_path_user}" + r"/*.json"):
+            logger.debug(f"加载用户词库: {file}")
+            try:
+                with open(file, "r") as f:
+                    for k, v in json.load(f).items():
+                        if k in word_dict:
+                            word_dict[k].extend(v)
+                        else:
+                            word_dict[k] = v
+            except Exception as e:
+                logger.error(f"加载词库失败: {file}: {e.__class__.__name__}: {e}")
+                continue
+    logger.debug(f"词库加载完成, 共加载词条: {len(word_dict)}")
+    return word_dict
+
+
+word_dict = _load_word_dict()
+ohayo_word = word_dict.get("早", ["早安", "早上好", "早上好呀", "早上好哦"])
+oyasumi_word = word_dict.get("晚安", ["晚安", "晚安呀", "晚安哦", "晚安喵"])
